fix(ML-609): move retry strategy to own folder
