+++ packages/askui-nodejs/src/execution/retry-strategies/exponential-retry-strategy.ts
@@ -1,4 +1,4 @@
-import { RetryStrategy } from '../retry-strategy';
+import { RetryStrategy } from './retry-strategy';
 
 /**
  * ExponentialRetryStrategy implements a retry strategy that uses an exponential backoff algorithm.
