+++ packages/askui-nodejs/src/execution/retry-strategies/linear-retry-strategy.ts
@@ -1,4 +1,4 @@
-import { RetryStrategy } from '../retry-strategy';
+import { RetryStrategy } from './retry-strategy';
 
 /**
  * LinearRetryStrategy implements a retry strategy that uses a linear backoff algorithm.
